requirements.txt
flask==2.0.1
pandas==1.3.3
numpy==1.21.2
tensorflow==2.6.0
scikit-learn==0.24.2
python-dotenv==0.19.0
flask-cors==3.0.10
requests==2.26.0
plotly==5.3.1
flask-bcrypt==1.0.1
pyjwt==2.1.0
